nitfinal/requirements.txt
gradio
langchain
langchain-community
langchain-groq
langchain-text-splitters
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
pypdf
pyttsx3
matplotlib
